requirements.txt
bleach==3.3.0
build==0.3.1.post1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
docutils==0.16
idna==2.10
importlib-metadata==3.9.1
jeepney==0.6.0
Jinja2==2.11.3
keyring==23.0.1
MarkupSafe==1.1.1
packaging==20.9
pep517==0.10.0
pkginfo==1.7.0
pyaml==20.4.0
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
PyYAML==5.4.1
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
SecretStorage==3.3.1
six==1.15.0
toml==0.10.2
tqdm==4.59.0
twine==3.4.1
urllib3==1.26.4
webencodings==0.5.1
zipp==3.4.1
